feature: Lua libraries/addons
Adds a new "require" function which will allow you load lua files in the "lua-libraries" folder. The purpose of this is to create addons, utils, or wrappers. The difference between this and the already existing addon api is that these addons are made in Lua and do not rely on another Java Plugin